Move most logic from Runner into VersionDataCollector (still need to update _do_translations)
